org settings: Add generalised function for tip-box for org settings.
This comit adds a `insert_tip_box` function and thus adds a
`organization-settings-tip` handlebars file to add a genaralised tip for
all organization settings pages. This further removes the code for tipbox
which was added in `popualte_auth_mmethods` functions, as it wasn''t making
sense there, making it more clear and readable.